refactor(db): Split block_builder.rs into directory module
Split block_builder.rs (1,155 lines) into 6 focused files:
- mod.rs: shared types (BlockResult, FieldHeadEntry), helper fns, re-exports
- write.rs: write_document_blocks(), write_delete_block()
- collection.rs: write_collection_block()
- build.rs: build_blocks_from_document(), build_block_from_document()
- read.rs: read_latest_composite_block()
- tests.rs: all test code
