--- crates/db/src/block_builder.rs
@@ -1,1155 +1,0 @@
-//! IPLD block builder for document mutations.
-//!
-//! Creates proper Block structures with CRDT delta payloads for P2P synchronization.
-//! Matches Go DefraDB's block format for wire compatibility.
-//!
-//! This module provides two main functions:
-//! - `build_blocks_from_document`: For P2P broadcast (uses external blockstore)
-//! - `write_document_blocks`: For FFI/local storage (uses transaction stores)
-
-use blockstore::Blockstore;
-use cid::Cid;
-use crypto::PrivateKey;
-use datastore::NamespaceView;
-use defra_core::block::{
-    Block, CollectionDeltaPayload, CompositeDeltaPayload, CounterDeltaPayload, CrdtDelta, DAGLink,
-    Encryption, LwwDeltaPayload, Signature, SignatureHeader, SignatureType,
-};
-use defra_core::encryption::EncryptionConfig;
-use defra_core::signing::SigningConfig;
-use document::{Document, NormalValue};
-use std::sync::Arc;
-use storage::corekv::Key;
-use storage::keys::headstore::{HeadstoreColKey, HeadstoreDocKey};
-
-fn encrypt_delta(plaintext: &[u8], key: &[u8; 32]) -> Result<Vec<u8>, String> {
-    let (ciphertext, _nonce) = crypto::encryption::aes::encrypt_aes(plaintext, key, &[], true)
-        .map_err(|e| format!("encryption failed: {}", e))?;
-    Ok(ciphertext)
-}
-
-/// Sign a block and store the signature as a separate IPLD block.
-///
-/// Matches Go's `signBlock()` in `internal/core/block/signing.go`:
-/// 1. Only signs first field blocks (priority <= 1) and composite blocks
-/// 2. Marshals block WITHOUT signature to get bytes to sign
-/// 3. Signs the bytes with the private key
-/// 4. Creates a Signature block with header (type + public key) and value
-/// 5. Stores the signature block in blockstore
-/// 6. Returns the signature block's CID
-///
-/// The caller must then set `block.signature = Some(sig_cid)` and re-serialize.
-async fn sign_block(
-    block: &Block,
-    signer: &SigningConfig,
-    blockstore: &NamespaceView,
-) -> Result<Option<Cid>, String> {
-    // Only sign first field blocks (priority <= 1) and composite blocks.
-    // Higher-priority field blocks are not signed — their integrity is
-    // guaranteed by the signature on the parent composite block.
-    let is_field = matches!(&block.delta, CrdtDelta::Lww(_) | CrdtDelta::Counter(_));
-    if is_field && block.delta.priority() > 1 {
-        return Ok(None);
-    }
-
-    // Serialize the block (without signature) to get the bytes to sign
-    let block_bytes = block
-        .to_dag_cbor()
-        .map_err(|e| format!("Failed to encode block for signing: {}", e))?;
-
-    // Determine signature type and sign
-    let (sig_type, sig_bytes) = match signer.key_type.as_str() {
-        "ed25519" => {
-            let private_key = crypto::Ed25519PrivateKey::from_bytes(&signer.private_key_bytes)
-                .map_err(|e| format!("Failed to load Ed25519 private key: {}", e))?;
-            let sig = private_key
-                .sign(&block_bytes)
-                .map_err(|e| format!("Failed to sign block: {}", e))?;
-            (SignatureType::EdDSA, sig)
-        }
-        "secp256k1" => {
-            let private_key = crypto::Secp256k1PrivateKey::from_bytes(&signer.private_key_bytes)
-                .map_err(|e| format!("Failed to load secp256k1 private key: {}", e))?;
-            let sig = private_key
-                .sign(&block_bytes)
-                .map_err(|e| format!("Failed to sign block: {}", e))?;
-            (SignatureType::ES256K, sig)
-        }
-        other => return Err(format!("Unsupported key type for signing: {}", other)),
-    };
-
-    // Create signature block.
-    // Go uses `[]byte(fullIdent.PublicKey().String())` for identity,
-    // which is the hex-encoded public key string as bytes.
-    let signature = Signature::new(
-        SignatureHeader::new(sig_type, signer.public_key_hex.as_bytes().to_vec()),
-        sig_bytes,
-    );
-
-    // Store signature block in blockstore and return its CID
-    let sig_cbor = signature
-        .to_dag_cbor()
-        .map_err(|e| format!("Failed to encode signature block: {}", e))?;
-    let sig_cid = signature
-        .generate_cid()
-        .map_err(|e| format!("Failed to generate signature CID: {}", e))?;
-
-    blockstore
-        .set(&sig_cid.to_bytes(), &sig_cbor)
-        .await
-        .map_err(|e| format!("Failed to store signature block: {}", e))?;
-
-    Ok(Some(sig_cid))
-}
-
-/// Result of building blocks from a document mutation.
-#[derive(Debug, Clone)]
-pub struct BlockResult {
-    /// The CID of the composite (root) block
-    pub cid: Cid,
-    /// The raw composite block bytes (DAG-CBOR encoded)
-    pub block: Vec<u8>,
-    /// The document ID
-    pub doc_id: String,
-    /// CIDs of all field blocks created
-    pub field_cids: Vec<Cid>,
-}
-
-/// Build IPLD blocks from a document for P2P sync.
-///
-/// This creates the proper Block structure that Go DefraDB expects:
-/// 1. An LWW field block for each document field
-/// 2. A Composite block linking all field blocks
-///
-/// # Arguments
-///
-/// * `doc` - The document to build blocks from
-/// * `schema_version_id` - The schema version ID for CRDT deltas
-/// * `blockstore` - Blockstore to store field blocks
-///
-/// # Returns
-///
-/// Returns `BlockResult` containing the composite block and metadata.
-pub async fn build_blocks_from_document<B: Blockstore>(
-    doc: &Document,
-    schema_version_id: &str,
-    blockstore: &Arc<B>,
-) -> Result<BlockResult, String> {
-    let doc_id = doc
-        .id()
-        .ok_or_else(|| "Document must have an ID".to_string())?;
-    let doc_id_str = doc_id.to_string();
-    let doc_id_bytes = doc_id_str.as_bytes().to_vec();
-
-    let mut field_links: Vec<DAGLink> = Vec::new();
-    let mut field_cids: Vec<Cid> = Vec::new();
-
-    // Create a block for each field (LWW or Counter depending on CRDT type)
-    for (field_name, field_value) in doc.values() {
-        // Skip _docID (stored as the block key, not a CRDT field).
-        // Other _-prefixed fields like _ownerID (foreign keys) DO get blocks.
-        if field_name == "_docID" {
-            continue;
-        }
-
-        // Encode the field value as CBOR
-        let value_bytes = encode_value_as_cbor(field_value.value())?;
-
-        // Check if this field is a Counter type
-        let is_counter = doc
-            .fields()
-            .get(field_name)
-            .map(|f| f.crdt_type().is_counter())
-            .unwrap_or(false);
-
-        // Create the appropriate delta based on CRDT type
-        let delta = if is_counter {
-            CrdtDelta::Counter(CounterDeltaPayload {
-                doc_id: doc_id_bytes.clone(),
-                field_name: field_name.clone(),
-                priority: 1, // Initial priority for new documents
-                nonce: 0,    // Go uses nonce=0 for initial creation (deterministic CID)
-                schema_version_id: schema_version_id.to_string(),
-                data: value_bytes,
-            })
-        } else {
-            CrdtDelta::Lww(LwwDeltaPayload {
-                doc_id: doc_id_bytes.clone(),
-                field_name: field_name.clone(),
-                priority: 1, // Initial priority for new documents
-                schema_version_id: schema_version_id.to_string(),
-                data: value_bytes,
-            })
-        };
-
-        // Create the field block
-        let field_block = Block::new(delta, vec![], vec![]);
-
-        // Serialize and generate CID
-        let field_block_bytes = field_block
-            .to_dag_cbor()
-            .map_err(|e| format!("Failed to encode field block: {}", e))?;
-        let field_cid = field_block
-            .generate_cid()
-            .map_err(|e| format!("Failed to generate field CID: {}", e))?;
-
-        // Store the field block in the blockstore
-        blockstore
-            .put(&field_cid, &field_block_bytes)
-            .await
-            .map_err(|e| format!("Failed to store field block: {}", e))?;
-
-        tracing::debug!(
-            field_name = %field_name,
-            cid = %field_cid,
-            is_counter = is_counter,
-            "Stored field block"
-        );
-
-        // Add link to composite
-        field_links.push(DAGLink::new(field_name.clone(), field_cid));
-        field_cids.push(field_cid);
-    }
-
-    // Create the Composite delta payload
-    let composite_payload = CompositeDeltaPayload {
-        doc_id: doc_id_bytes,
-        schema_version_id: schema_version_id.to_string(),
-        priority: 1, // Initial priority
-        status: 1,   // Active document
-    };
-
-    // Create the composite block with links to all field blocks
-    let composite_block = Block::new(CrdtDelta::Composite(composite_payload), vec![], field_links);
-
-    // Serialize the composite block
-    let composite_bytes = composite_block
-        .to_dag_cbor()
-        .map_err(|e| format!("Failed to encode composite block: {}", e))?;
-    let composite_cid = composite_block
-        .generate_cid()
-        .map_err(|e| format!("Failed to generate composite CID: {}", e))?;
-
-    // Store the composite block
-    blockstore
-        .put(&composite_cid, &composite_bytes)
-        .await
-        .map_err(|e| format!("Failed to store composite block: {}", e))?;
-
-    tracing::info!(
-        doc_id = %doc_id_str,
-        cid = %composite_cid,
-        field_count = field_cids.len(),
-        "Built composite block with field links"
-    );
-
-    Ok(BlockResult {
-        cid: composite_cid,
-        block: composite_bytes,
-        doc_id: doc_id_str,
-        field_cids,
-    })
-}
-
-/// Encode a NormalValue as CBOR bytes.
-fn encode_value_as_cbor(value: &NormalValue) -> Result<Vec<u8>, String> {
-    let mut bytes = Vec::new();
-    ciborium::into_writer(value, &mut bytes)
-        .map_err(|e| format!("Failed to encode value as CBOR: {}", e))?;
-    Ok(bytes)
-}
-
-/// Encode a priority as a varint (matching Go's binary.PutUvarint).
-fn encode_priority_varint(priority: u64) -> Vec<u8> {
-    let mut buf = Vec::with_capacity(10); // Max varint64 is 10 bytes
-    let mut n = priority;
-    while n >= 0x80 {
-        buf.push((n as u8) | 0x80);
-        n >>= 7;
-    }
-    buf.push(n as u8);
-    buf
-}
-
-/// Decode a varint to priority (matching Go's binary.Uvarint).
-fn decode_priority_varint(buf: &[u8]) -> u64 {
-    let mut n: u64 = 0;
-    let mut shift: u32 = 0;
-    for &byte in buf {
-        if shift >= 64 {
-            return 0; // Overflow protection
-        }
-        n |= ((byte & 0x7f) as u64) << shift;
-        if byte < 0x80 {
-            return n;
-        }
-        shift += 7;
-    }
-    n
-}
-
-/// A single head entry for a document field.
-struct FieldHeadEntry {
-    /// The CID of the head
-    cid: Cid,
-    /// The full key (for deletion when replacing)
-    key: Vec<u8>,
-}
-
-/// Get all existing heads for a specific field of a document.
-///
-/// During concurrent P2P updates, a field can have multiple heads (branches).
-/// Returns all current head CIDs for the field, sorted by CID string
-/// representation to match Go's deterministic head ordering.
-async fn get_all_field_heads(
-    headstore: &NamespaceView,
-    doc_id: &str,
-    field_id: &str,
-) -> Result<Vec<FieldHeadEntry>, String> {
-    use storage::corekv::IterOptions;
-
-    let prefix = HeadstoreDocKey::field_prefix(doc_id, field_id);
-    let opts = IterOptions::new().with_prefix(prefix);
-
-    let mut iter = headstore
-        .iterator(opts)
-        .await
-        .map_err(|e| format!("Failed to create headstore iterator: {}", e))?;
-
-    let mut entries = Vec::new();
-    while let Some(kv_pair) = iter
-        .next()
-        .await
-        .map_err(|e| format!("Failed to iterate headstore: {}", e))?
-    {
-        // Parse CID from key: /d/{doc_id}/{field_id}/{cid}
-        let key_str = String::from_utf8_lossy(&kv_pair.key);
-        let parts: Vec<&str> = key_str.split('/').collect();
-        if let Some(cid_str) = parts.last() {
-            if let Ok(cid) = cid_str.parse::<Cid>() {
-                entries.push(FieldHeadEntry {
-                    cid,
-                    key: kv_pair.key.clone(),
-                });
-            }
-        }
-    }
-
-    // Sort by CID string representation to match Go's Block.New() sorting
-    entries.sort_by(|a, b| a.cid.to_string().cmp(&b.cid.to_string()));
-    Ok(entries)
-}
-
-/// Get the maximum priority from existing heads for a document.
-///
-/// Scans the headstore for all existing heads of the given document
-/// and returns the maximum priority found. Returns 0 if no heads exist.
-async fn get_max_priority(headstore: &NamespaceView, doc_id: &str) -> Result<u64, String> {
-    use storage::corekv::IterOptions;
-
-    let prefix = HeadstoreDocKey::document_prefix(doc_id);
-    let opts = IterOptions::new().with_prefix(prefix);
-
-    let mut iter = headstore
-        .iterator(opts)
-        .await
-        .map_err(|e| format!("Failed to create headstore iterator: {}", e))?;
-
-    let mut max_priority: u64 = 0;
-    while let Some(kv_pair) = iter
-        .next()
-        .await
-        .map_err(|e| format!("Failed to iterate headstore: {}", e))?
-    {
-        let priority = decode_priority_varint(&kv_pair.value);
-        if priority > max_priority {
-            max_priority = priority;
-        }
-    }
-
-    Ok(max_priority)
-}
-
-/// Write document blocks to blockstore and heads to headstore.
-///
-/// This function creates CRDT blocks for a document and writes:
-/// 1. LWW field blocks to blockstore
-/// 2. Composite block to blockstore
-/// 3. Head CIDs to headstore (for _commits queries)
-///
-/// This matches Go DefraDB's ProcessBlock → updateHeads flow.
-///
-/// # Arguments
-///
-/// * `blockstore` - Transaction blockstore view
-/// * `headstore` - Transaction headstore view
-/// * `doc` - The document to build blocks from
-/// * `schema_version_id` - The schema version ID for CRDT deltas
-/// * `modified_fields` - Optional set of field names that were modified.
-///   If Some, only these fields will have new blocks created.
-///   If None (for create operations), all fields get new blocks.
-///
-/// # Returns
-///
-/// Returns `BlockResult` containing the composite block and metadata.
-pub async fn write_document_blocks(
-    blockstore: &NamespaceView,
-    headstore: &NamespaceView,
-    doc: &Document,
-    schema_version_id: &str,
-    modified_fields: Option<&std::collections::HashSet<String>>,
-    encryption_config: Option<&EncryptionConfig>,
-    signing_config: Option<&SigningConfig>,
-) -> Result<BlockResult, String> {
-    let doc_id = doc
-        .id()
-        .ok_or_else(|| "Document must have an ID".to_string())?;
-    let doc_id_str = doc_id.to_string();
-    let doc_id_bytes = doc_id_str.as_bytes().to_vec();
-
-    let mut field_links: Vec<DAGLink> = Vec::new();
-    let mut field_cids: Vec<Cid> = Vec::new();
-
-    // For creates (modified_fields=None), priority is always 1 and no heads exist.
-    // For updates, scan headstore for max priority (matches Go behavior).
-    let is_create = modified_fields.is_none();
-    let priority: u64 = if is_create {
-        1
-    } else {
-        get_max_priority(headstore, &doc_id_str).await? + 1
-    };
-
-    // Create LWW blocks for each field
-    // For updates with modified_fields, only create blocks for changed fields.
-    // For unchanged fields, use their existing head CID in the composite links.
-    for (field_name, field_value) in doc.values() {
-        // Skip _docID (stored as the block key, not a CRDT field).
-        // Other _-prefixed fields like _ownerID (foreign keys) DO get blocks.
-        if field_name == "_docID" {
-            continue;
-        }
-
-        // Check if this field should have a new block created
-        let should_create_block = match modified_fields {
-            None => true, // Create operation: all fields get new blocks
-            Some(fields) => fields.contains(field_name), // Update: only modified fields
-        };
-
-        // For creates, no heads exist yet — skip the headstore scan.
-        let field_head_entries = if is_create {
-            Vec::new()
-        } else {
-            get_all_field_heads(headstore, &doc_id_str, field_name).await?
-        };
-
-        if should_create_block {
-            // Create new block for this field (LWW or Counter depending on CRDT type)
-            let heads: Vec<Cid> = field_head_entries.iter().map(|h| h.cid).collect();
-
-            // For counter fields during updates, use the raw increment delta
-            // (not the accumulated value) to match Go behavior.
-            let cbor_value = if let Some(delta) = doc.get_counter_delta(field_name) {
-                delta
-            } else {
-                field_value.value()
-            };
-
-            // Encode the field value as CBOR
-            let value_bytes = encode_value_as_cbor(cbor_value)?;
-
-            // Encrypt delta and create Encryption metadata block if configured
-            let (value_bytes, encryption_cid) = if let Some(enc) = encryption_config {
-                if enc.should_encrypt_field(field_name) {
-                    let key = enc.derive_key(field_name, &doc_id_str);
-                    let encrypted = encrypt_delta(&value_bytes, &key)?;
-
-                    tracing::debug!(
-                        field = %field_name,
-                        ciphertext_len = encrypted.len(),
-                        "Encrypted field delta"
-                    );
-
-                    // Create Encryption metadata block (matches Go's block/encryption.go)
-                    let is_field_level = enc.encrypt_fields.iter().any(|f| f == field_name);
-                    let enc_block = Encryption {
-                        doc_id: doc_id_bytes.clone(),
-                        field_name: if is_field_level {
-                            Some(field_name.clone())
-                        } else {
-                            None
-                        },
-                        key: key.to_vec(),
-                    };
-                    let enc_cid = enc_block
-                        .generate_cid()
-                        .map_err(|e| format!("Failed to generate encryption CID: {}", e))?;
-                    let enc_bytes = enc_block
-                        .to_dag_cbor()
-                        .map_err(|e| format!("Failed to encode encryption block: {}", e))?;
-                    blockstore
-                        .set(&enc_cid.to_bytes(), &enc_bytes)
-                        .await
-                        .map_err(|e| format!("Failed to store encryption block: {}", e))?;
-
-                    (encrypted, Some(enc_cid))
-                } else {
-                    (value_bytes, None)
-                }
-            } else {
-                (value_bytes, None)
-            };
-
-            // Check if this field is a Counter type.
-            // For documents loaded from storage, the CRDT type may default to LWW
-            // since storage doesn't preserve schema annotations. Fall back to checking
-            // whether a counter delta was set (only counter fields get counter deltas).
-            let is_counter = doc
-                .fields()
-                .get(field_name)
-                .map(|f| f.crdt_type().is_counter())
-                .unwrap_or(false)
-                || doc.get_counter_delta(field_name).is_some();
-
-            // For Counter fields, use nonce=0 for initial creation (heads empty),
-            // random nonce for updates (heads non-empty) - matches Go behavior
-            let nonce: i64 = if is_counter && !heads.is_empty() {
-                // Update operation: use random nonce
-                rand::random()
-            } else {
-                // Initial creation: use deterministic nonce=0
-                0
-            };
-
-            // Create the appropriate delta based on CRDT type
-            let delta = if is_counter {
-                CrdtDelta::Counter(CounterDeltaPayload {
-                    doc_id: doc_id_bytes.clone(),
-                    field_name: field_name.clone(),
-                    priority,
-                    nonce,
-                    schema_version_id: schema_version_id.to_string(),
-                    data: value_bytes,
-                })
-            } else {
-                CrdtDelta::Lww(LwwDeltaPayload {
-                    doc_id: doc_id_bytes.clone(),
-                    field_name: field_name.clone(),
-                    priority,
-                    schema_version_id: schema_version_id.to_string(),
-                    data: value_bytes,
-                })
-            };
-
-            // Create the field block with heads linking to previous version
-            // Include encryption CID if this field was encrypted
-            let mut field_block =
-                Block::new_with_options(delta, heads.clone(), vec![], encryption_cid, None);
-
-            // Sign the block if a signer is available.
-            // Go's signBlock: sign the block bytes (without signature), store
-            // the Signature as a separate block, then set block.signature = sig_cid.
-            if let Some(signer) = signing_config {
-                tracing::debug!(field = %field_name, priority, "Signing field block");
-                match sign_block(&field_block, signer, blockstore).await {
-                    Ok(Some(sig_cid)) => {
-                        tracing::debug!(field = %field_name, sig_cid = %sig_cid, "Field block signed");
-                        field_block.signature = Some(sig_cid);
-                    }
-                    Ok(None) => {
-                        tracing::debug!(field = %field_name, priority, "Skipped signing (priority > 1)");
-                    }
-                    Err(e) => {
-                        tracing::debug!(field = %field_name, error = %e, "Failed to sign field block");
-                        return Err(e);
-                    }
-                }
-            } else {
-                tracing::debug!(field = %field_name, "No signing config for field");
-            }
-
-            // Serialize and generate CID (includes signature CID if signed)
-            let field_block_bytes = field_block
-                .to_dag_cbor()
-                .map_err(|e| format!("Failed to encode field block: {}", e))?;
-            let field_cid = field_block
-                .generate_cid()
-                .map_err(|e| format!("Failed to generate field CID: {}", e))?;
-
-            // Debug: verify roundtrip of signature through serialization
-            tracing::debug!(
-                field = %field_name,
-                cid = %field_cid,
-                bytes_len = field_block_bytes.len(),
-                signature = ?field_block.signature,
-                "Storing field block"
-            );
-            if tracing::enabled!(tracing::Level::DEBUG) {
-                match Block::from_dag_cbor(&field_block_bytes) {
-                    Ok(b) => tracing::debug!(signature = ?b.signature, "Signature roundtrip OK"),
-                    Err(e) => tracing::debug!(error = %e, "Signature roundtrip failed"),
-                }
-            }
-
-            // Store the field block in blockstore
-            blockstore
-                .set(&field_cid.to_bytes(), &field_block_bytes)
-                .await
-                .map_err(|e| format!("Failed to store field block: {}", e))?;
-
-            // Delete all old head entries (replace all branches with single new head)
-            for old_head in &field_head_entries {
-                headstore
-                    .delete(&old_head.key)
-                    .await
-                    .map_err(|e| format!("Failed to delete old field head: {}", e))?;
-            }
-
-            // Write new head CID to headstore: /d/{doc_id}/{field_name}/{cid} → priority
-            let head_key = HeadstoreDocKey::new(&doc_id_str, field_name, field_cid);
-            let priority_bytes = encode_priority_varint(priority);
-            headstore
-                .set(&head_key.bytes(), &priority_bytes)
-                .await
-                .map_err(|e| format!("Failed to write field head: {}", e))?;
-
-            tracing::debug!(
-                field_name = %field_name,
-                cid = %field_cid,
-                is_counter = is_counter,
-                has_prev_head = !heads.is_empty(),
-                "Stored field block and head"
-            );
-
-            // Add link to composite - only new blocks get linked
-            field_links.push(DAGLink::new(field_name.clone(), field_cid));
-            field_cids.push(field_cid);
-        }
-        // Note: Unchanged fields are NOT added to composite links.
-        // Go only includes newly created field blocks in the composite's links array.
-    }
-
-    // For creates, no composite heads exist yet — skip the headstore scan.
-    let (composite_head_entries, composite_heads) = if is_create {
-        (Vec::new(), Vec::new())
-    } else {
-        let entries = get_all_field_heads(headstore, &doc_id_str, "C").await?;
-        let heads: Vec<Cid> = entries.iter().map(|h| h.cid).collect();
-        (entries, heads)
-    };
-
-    // Create the Composite delta payload
-    let composite_payload = CompositeDeltaPayload {
-        doc_id: doc_id_bytes,
-        schema_version_id: schema_version_id.to_string(),
-        priority,
-        status: 1, // Active document
-    };
-
-    // For doc-level encryption, the composite block also carries an encryption
-    // CID link (even though its data is not encrypted). Go's AddDelta always calls
-    // determineBlockEncryption, which returns an encryption block for doc-level
-    // encryption regardless of delta type. encryptBlock() then skips actual
-    // encryption for composites but the Encryption link is still set on the block.
-    let composite_encryption_cid = if let Some(enc) = encryption_config {
-        if enc.encrypt_doc {
-            let key = enc.derive_key("", &doc_id_str); // doc-level: empty field name
-            let enc_block = Encryption {
-                doc_id: doc_id_str.as_bytes().to_vec(),
-                field_name: None,
-                key: key.to_vec(),
-            };
-            let enc_cid = enc_block
-                .generate_cid()
-                .map_err(|e| format!("Failed to generate composite encryption CID: {}", e))?;
-            let enc_bytes = enc_block
-                .to_dag_cbor()
-                .map_err(|e| format!("Failed to encode composite encryption block: {}", e))?;
-            blockstore
-                .set(&enc_cid.to_bytes(), &enc_bytes)
-                .await
-                .map_err(|e| format!("Failed to store composite encryption block: {}", e))?;
-            Some(enc_cid)
-        } else {
-            None
-        }
-    } else {
-        None
-    };
-
-    // Create the composite block with heads linking to previous version
-    let mut composite_block = Block::new_with_options(
-        CrdtDelta::Composite(composite_payload),
-        composite_heads,
-        field_links,
-        composite_encryption_cid,
-        None,
-    );
-
-    // Sign the composite block if a signer is available
-    if let Some(signer) = signing_config {
-        if let Some(sig_cid) = sign_block(&composite_block, signer, blockstore).await? {
-            composite_block.signature = Some(sig_cid);
-        }
-    }
-
-    // Serialize the composite block (includes signature CID if signed)
-    let composite_bytes = composite_block
-        .to_dag_cbor()
-        .map_err(|e| format!("Failed to encode composite block: {}", e))?;
-    let composite_cid = composite_block
-        .generate_cid()
-        .map_err(|e| format!("Failed to generate composite CID: {}", e))?;
-
-    // Store the composite block in blockstore
-    blockstore
-        .set(&composite_cid.to_bytes(), &composite_bytes)
-        .await
-        .map_err(|e| format!("Failed to store composite block: {}", e))?;
-
-    // Delete all old composite head entries (replace all branches with single new head)
-    for old_head in &composite_head_entries {
-        headstore
-            .delete(&old_head.key)
-            .await
-            .map_err(|e| format!("Failed to delete old composite head: {}", e))?;
-    }
-
-    // Write new composite head to headstore: /d/{doc_id}/C/{cid} → priority
-    let composite_head_key = HeadstoreDocKey::new(&doc_id_str, "C", composite_cid);
-    let priority_bytes = encode_priority_varint(priority);
-    headstore
-        .set(&composite_head_key.bytes(), &priority_bytes)
-        .await
-        .map_err(|e| format!("Failed to write composite head: {}", e))?;
-
-    tracing::info!(
-        doc_id = %doc_id_str,
-        cid = %composite_cid,
-        field_count = field_cids.len(),
-        prev_heads = composite_head_entries.len(),
-        "Built composite block with field links and wrote heads"
-    );
-
-    Ok(BlockResult {
-        cid: composite_cid,
-        block: composite_bytes,
-        doc_id: doc_id_str,
-        field_cids,
-    })
-}
-
-/// Write a delete block (composite with status=2) to blockstore and heads to headstore.
-///
-/// Delete operations in Go DefraDB create a composite block with status=2 (deleted),
-/// no field-level blocks, and heads pointing to the previous composite head.
-/// This matches Go's `crdt.DocComposite.DeleteDelta()` → `coreblock.AddDelta()` flow.
-pub async fn write_delete_block(
-    blockstore: &NamespaceView,
-    headstore: &NamespaceView,
-    doc_id: &str,
-    schema_version_id: &str,
-    signing_config: Option<&SigningConfig>,
-) -> Result<BlockResult, String> {
-    let doc_id_bytes = doc_id.as_bytes().to_vec();
-
-    // Get max existing priority and increment by 1
-    let max_priority = get_max_priority(headstore, doc_id).await?;
-    let priority: u64 = max_priority + 1;
-
-    // Get all existing composite heads to build proper DAG links
-    let composite_head_entries = get_all_field_heads(headstore, doc_id, "C").await?;
-    let composite_heads: Vec<Cid> = composite_head_entries.iter().map(|h| h.cid).collect();
-
-    // Create the Composite delta payload with status=2 (deleted)
-    let composite_payload = CompositeDeltaPayload {
-        doc_id: doc_id_bytes,
-        schema_version_id: schema_version_id.to_string(),
-        priority,
-        status: 2, // Deleted document
-    };
-
-    // Create the composite block with no field links (deletes have no field blocks)
-    let mut composite_block = Block::new(
-        CrdtDelta::Composite(composite_payload),
-        composite_heads,
-        vec![], // No field links for delete
-    );
-
-    // Sign the delete composite block if a signer is available
-    if let Some(signer) = signing_config {
-        if let Some(sig_cid) = sign_block(&composite_block, signer, blockstore).await? {
-            composite_block.signature = Some(sig_cid);
-        }
-    }
-
-    // Serialize the composite block
-    let composite_bytes = composite_block
-        .to_dag_cbor()
-        .map_err(|e| format!("Failed to encode delete composite block: {}", e))?;
-    let composite_cid = composite_block
-        .generate_cid()
-        .map_err(|e| format!("Failed to generate delete composite CID: {}", e))?;
-
-    // Store the composite block in blockstore
-    blockstore
-        .set(&composite_cid.to_bytes(), &composite_bytes)
-        .await
-        .map_err(|e| format!("Failed to store delete composite block: {}", e))?;
-
-    // Delete all old composite head entries (replace all branches with single new head)
-    for old_head in &composite_head_entries {
-        headstore
-            .delete(&old_head.key)
-            .await
-            .map_err(|e| format!("Failed to delete old composite head: {}", e))?;
-    }
-
-    // Write new composite head to headstore: /d/{doc_id}/C/{cid} → priority
-    let composite_head_key = HeadstoreDocKey::new(doc_id, "C", composite_cid);
-    let priority_bytes = encode_priority_varint(priority);
-    headstore
-        .set(&composite_head_key.bytes(), &priority_bytes)
-        .await
-        .map_err(|e| format!("Failed to write delete composite head: {}", e))?;
-
-    tracing::info!(
-        doc_id = %doc_id,
-        cid = %composite_cid,
-        priority = priority,
-        "Built delete composite block (status=2)"
-    );
-
-    Ok(BlockResult {
-        cid: composite_cid,
-        block: composite_bytes,
-        doc_id: doc_id.to_string(),
-        field_cids: vec![],
-    })
-}
-
-/// Write a collection-level block for branchable collections.
-///
-/// Branchable collections maintain a separate DAG at the collection level,
-/// tracking all document operations. Each mutation creates a collection block
-/// that links to the document's composite block.
-///
-/// This matches Go's `crdt.NewCollection()` → `coreblock.AddDelta()` flow.
-pub async fn write_collection_block(
-    blockstore: &NamespaceView,
-    headstore: &NamespaceView,
-    collection_short_id: u32,
-    schema_version_id: &str,
-    doc_composite_cid: Cid,
-    signing_config: Option<&SigningConfig>,
-) -> Result<(Cid, Vec<u8>), String> {
-    use storage::corekv::IterOptions;
-
-    // Get existing collection head (if any)
-    let col_prefix = HeadstoreColKey::collection_prefix(collection_short_id);
-    let opts = IterOptions::new().with_prefix(col_prefix);
-
-    let mut iter = headstore
-        .iterator(opts)
-        .await
-        .map_err(|e| format!("Failed to create collection headstore iterator: {}", e))?;
-
-    let mut col_heads: Vec<Cid> = Vec::new();
-    let mut max_priority: u64 = 0;
-    let mut old_head_keys: Vec<Vec<u8>> = Vec::new();
-
-    while let Some(kv_pair) = iter
-        .next()
-        .await
-        .map_err(|e| format!("Failed to iterate collection headstore: {}", e))?
-    {
-        let priority = decode_priority_varint(&kv_pair.value);
-        if priority > max_priority {
-            max_priority = priority;
-        }
-        // Parse CID from key: /c/{collection_id}/{cid}
-        let key_str = String::from_utf8_lossy(&kv_pair.key);
-        let parts: Vec<&str> = key_str.split('/').collect();
-        if let Some(cid_str) = parts.last() {
-            if let Ok(cid) = cid_str.parse::<Cid>() {
-                col_heads.push(cid);
-            }
-        }
-        old_head_keys.push(kv_pair.key.clone());
-    }
-
-    let priority: u64 = max_priority + 1;
-
-    // Sort heads by CID string representation to match Go's Block.New() sorting
-    col_heads.sort_by_key(|a| a.to_string());
-
-    // Create the Collection delta payload
-    let collection_payload = CollectionDeltaPayload {
-        schema_version_id: schema_version_id.to_string(),
-        priority,
-    };
-
-    // The collection block links to the document composite block
-    // Go uses empty string for link name here, fieldName comes from linked block's delta
-    let links = vec![DAGLink::new(String::new(), doc_composite_cid)];
-
-    // Create the collection block
-    let mut collection_block =
-        Block::new(CrdtDelta::Collection(collection_payload), col_heads, links);
-
-    // Sign the collection block if a signer is available
-    if let Some(signer) = signing_config {
-        if let Some(sig_cid) = sign_block(&collection_block, signer, blockstore).await? {
-            collection_block.signature = Some(sig_cid);
-        }
-    }
-
-    // Serialize and generate CID
-    let collection_bytes = collection_block
-        .to_dag_cbor()
-        .map_err(|e| format!("Failed to encode collection block: {}", e))?;
-    let collection_cid = collection_block
-        .generate_cid()
-        .map_err(|e| format!("Failed to generate collection CID: {}", e))?;
-
-    // Store the collection block in blockstore
-    blockstore
-        .set(&collection_cid.to_bytes(), &collection_bytes)
-        .await
-        .map_err(|e| format!("Failed to store collection block: {}", e))?;
-
-    // Delete old collection head entries
-    for old_key in old_head_keys {
-        headstore
-            .delete(&old_key)
-            .await
-            .map_err(|e| format!("Failed to delete old collection head: {}", e))?;
-    }
-
-    // Write new collection head: /c/{collection_id}/{cid} → priority
-    let col_head_key = HeadstoreColKey::new(collection_short_id, collection_cid);
-    let priority_bytes = encode_priority_varint(priority);
-    headstore
-        .set(&col_head_key.bytes(), &priority_bytes)
-        .await
-        .map_err(|e| format!("Failed to write collection head: {}", e))?;
-
-    tracing::info!(
-        collection_id = collection_short_id,
-        cid = %collection_cid,
-        priority = priority,
-        doc_composite_cid = %doc_composite_cid,
-        "Built collection block for branchable collection"
-    );
-
-    Ok((collection_cid, collection_bytes))
-}
-
-// === Legacy function for backwards compatibility ===
-
-/// Build an IPLD block from a document (legacy - DO NOT USE for P2P).
-///
-/// Read the latest composite block for a document from the committed store.
-///
-/// After a mutation (create/update/delete), the composite block and its CID
-/// are written to the blockstore and headstore as part of the transaction.
-/// This function reads the committed composite head CID from the headstore,
-/// then fetches the corresponding block bytes from the blockstore.
-///
-/// Use this instead of `build_blocks_from_document` for P2P broadcast after
-/// updates/deletes, since `build_blocks_from_document` recreates blocks from
-/// scratch with wrong priority/heads.
-pub async fn read_latest_composite_block<S: storage::corekv::Store>(
-    db: &crate::database::DB<S>,
-    doc_id: &str,
-) -> Result<BlockResult, String> {
-    let txn = db
-        .new_txn(true)
-        .await
-        .map_err(|e| format!("Failed to create read txn: {}", e))?;
-
-    let headstore = txn
-        .headstore()
-        .map_err(|e| format!("Failed to get headstore: {}", e))?;
-
-    let composite_heads = get_all_field_heads(&headstore, doc_id, "C").await?;
-
-    let composite_cid = composite_heads
-        .first()
-        .map(|entry| entry.cid)
-        .ok_or_else(|| format!("No composite head found for doc {}", doc_id))?;
-
-    let blockstore = txn
-        .blockstore()
-        .map_err(|e| format!("Failed to get blockstore: {}", e))?;
-
-    let block_bytes = blockstore
-        .get(&composite_cid.to_bytes())
-        .await
-        .map_err(|e| format!("Failed to read composite block: {}", e))?
-        .ok_or_else(|| format!("Composite block not found for CID {}", composite_cid))?;
-
-    let _ = txn.force_discard();
-
-    Ok(BlockResult {
-        cid: composite_cid,
-        block: block_bytes,
-        doc_id: doc_id.to_string(),
-        field_cids: vec![],
-    })
-}
-
-/// **WARNING**: This creates raw document CBOR, not a proper Block structure.
-/// Go DefraDB cannot parse this format. Use `build_blocks_from_document` instead.
-///
-/// This function is kept for backward compatibility with non-P2P code paths.
-#[deprecated(since = "0.1.0", note = "Use build_blocks_from_document for P2P sync")]
-pub fn build_block_from_document(doc: &Document) -> Result<BlockResult, String> {
-    use multihash::MultihashGeneric;
-    use sha2::{Digest, Sha256};
-
-    const DAG_CBOR_CODEC: u64 = 0x71;
-    const SHA2_256_CODE: u64 = 0x12;
-
-    let doc_id = doc
-        .id()
-        .ok_or_else(|| "Document must have an ID".to_string())?
-        .to_string();
-
-    // Encode document as CBOR (raw - NOT a Block structure)
-    let block = doc
-        .to_cbor()
-        .map_err(|e| format!("Failed to encode document: {}", e))?;
-
-    // Create CID from block
-    let mut hasher = Sha256::new();
-    hasher.update(&block);
-    let digest = hasher.finalize();
-
-    let mh = MultihashGeneric::<64>::wrap(SHA2_256_CODE, &digest)
-        .map_err(|e| format!("Failed to create multihash: {}", e))?;
-
-    let cid = Cid::new_v1(DAG_CBOR_CODEC, mh);
-
-    Ok(BlockResult {
-        cid,
-        block,
-        doc_id,
-        field_cids: vec![],
-    })
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use blockstore::DefraBlockstore;
-    use storage::backends::MemoryStore;
-
-    fn make_test_blockstore() -> Arc<DefraBlockstore<MemoryStore>> {
-        let store = Arc::new(MemoryStore::new());
-        Arc::new(DefraBlockstore::new(store, false))
-    }
-
-    #[tokio::test]
-    async fn test_build_blocks_creates_proper_structure() {
-        let mut doc = Document::new();
-        doc.generate_and_set_doc_id().unwrap();
-        doc.set("name", NormalValue::String("Alice".to_string()));
-        doc.set("age", NormalValue::Int(30));
-
-        let blockstore = make_test_blockstore();
-        let schema_version_id = "bafyreihsneodeja4lfer5puptim3lkwvketyckrmkhfpgxm67ch5wenjwq";
-
-        let result = build_blocks_from_document(&doc, schema_version_id, &blockstore)
-            .await
-            .unwrap();
-
-        // Should have created 2 field blocks (name, age)
-        assert_eq!(result.field_cids.len(), 2);
-        assert!(!result.doc_id.is_empty());
-
-        // Composite block should be in blockstore
-        let stored = blockstore.get(&result.cid).await.unwrap();
-        assert!(stored.is_some());
-
-        // Each field block should be in blockstore
-        for field_cid in &result.field_cids {
-            let stored = blockstore.get(field_cid).await.unwrap();
-            assert!(stored.is_some());
-        }
-    }
-
-    #[tokio::test]
-    async fn test_build_blocks_requires_doc_id() {
-        let doc = Document::new();
-        let blockstore = make_test_blockstore();
-
-        let result = build_blocks_from_document(&doc, "schema-v1", &blockstore).await;
-        assert!(result.is_err());
-        assert!(result.unwrap_err().contains("must have an ID"));
-    }
-
-    #[tokio::test]
-    async fn test_field_block_contains_lww_delta() {
-        let mut doc = Document::new();
-        doc.generate_and_set_doc_id().unwrap();
-        doc.set("name", NormalValue::String("Bob".to_string()));
-
-        let blockstore = make_test_blockstore();
-        let schema_version_id = "schema-v1";
-
-        let result = build_blocks_from_document(&doc, schema_version_id, &blockstore)
-            .await
-            .unwrap();
-
-        // Get the field block
-        let field_cid = &result.field_cids[0];
-        let field_bytes = blockstore.get(field_cid).await.unwrap().unwrap();
-
-        // Decode and verify it's an LWW block
-        let field_block = Block::from_dag_cbor(&field_bytes).unwrap();
-        match &field_block.delta {
-            CrdtDelta::Lww(payload) => {
-                assert_eq!(payload.field_name, "name");
-                assert_eq!(payload.schema_version_id, schema_version_id);
-                assert_eq!(payload.priority, 1);
-            }
-            _ => panic!("Expected LWW delta"),
-        }
-    }
-
-    #[tokio::test]
-    async fn test_composite_block_has_field_links() {
-        let mut doc = Document::new();
-        doc.generate_and_set_doc_id().unwrap();
-        doc.set("name", NormalValue::String("Charlie".to_string()));
-        doc.set("age", NormalValue::Int(25));
-
-        let blockstore = make_test_blockstore();
-
-        let result = build_blocks_from_document(&doc, "schema-v1", &blockstore)
-            .await
-            .unwrap();
-
-        // Decode the composite block
-        let composite_block = Block::from_dag_cbor(&result.block).unwrap();
-
-        // Verify it's a Composite delta
-        match &composite_block.delta {
-            CrdtDelta::Composite(payload) => {
-                assert_eq!(payload.status, 1); // Active
-                assert_eq!(payload.priority, 1);
-            }
-            _ => panic!("Expected Composite delta"),
-        }
-
-        // Verify links to field blocks
-        let links = composite_block.links.as_ref().expect("Should have links");
-        assert_eq!(links.len(), 2);
-
-        // Links should reference field CIDs
-        let link_cids: Vec<Cid> = links.iter().map(|l| l.link).collect();
-        for field_cid in &result.field_cids {
-            assert!(link_cids.contains(field_cid));
-        }
-    }
-}
